Student Pseudonym Updates
The student pseudonym plugin now is now exclusively focused on changing the username, and only allows a student to do so once.

+++ class-blogs/ClassBlogs/Plugins/StudentPseudonym.php
@@ -20,6 +20,17 @@ ClassBlogs::require_cb_file( 'WordPress.php' );
  */
 class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 {
+	/**
+	 * The default options for the plugin.
+	 *
+	 * @access protected
+	 * @var array
+	 * @since 0.4
+	 */
+	protected $default_options = array(
+		'changed_users' => array()
+	);
+
 	/** Registers hooks to add the student pseudonym admin page. */
 	public function __construct()
 	{
@@ -28,16 +39,17 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 	}
 
 	/**
-	 * Register hooks to show the admin page when on a student blog.
-	 *
-	 * The pseudonym page will be only available when running in multisite mode.
+	 * Register hooks to show the admin page when a student who has not yet
+	 * changed their username is logged in to the admin side.
 	 *
 	 * @access private
 	 * @since 0.3
 	 */
 	public function _register_hooks()
 	{
-		if ( ClassBlogs_Utils::is_multisite() && ClassBlogs_Utils::on_student_blog_admin() ) {
+		$current_user = wp_get_current_user();
+		if ( $current_user->ID > 0 && ClassBlogs_Utils::on_student_blog_admin() &&
+		     ! array_key_exists( $current_user->user_login, $this->get_option( 'changed_users' ) ) ) {
 			add_action( 'admin_menu', array( $this, '_add_admin_page' ) );
 		}
 	}
@@ -51,9 +63,9 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 	public function _add_admin_page( $admin )
 	{
 		add_users_page(
-			__( 'Use A Pseudonym', 'classblogs' ),
-			__( 'Use A Pseudonym', 'classblogs' ),
-			'manage_options',
+			__( 'Change Username', 'classblogs' ),
+			__( 'Change Username', 'classblogs' ),
+			'edit_posts',
 			$this->get_uid(),
 			array( $this, '_admin_page' ) );
 	}
@@ -68,8 +80,6 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 	{
 		// Get information on the user and their blog URL
 		$current_user = wp_get_current_user();
-		$first_name = $current_user->user_firstname;
-		$last_name = $current_user->user_lastname;
 		$username = $current_user->user_login;
 		$blog_url = home_url();
 
@@ -81,27 +91,22 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 			// does not conflict with an existing user or blog
 			check_admin_referer( $this->get_uid() );
 			$username = ClassBlogs_Utils::sanitize_user_input( $_POST['new_username'] );
-			$first_name = ClassBlogs_Utils::sanitize_user_input( $_POST['new_first_name'] );
-			$last_name = ClassBlogs_Utils::sanitize_user_input( $_POST['new_last_name'] );
 			$username_valid = $this->_validate_username( $username );
 			if ( $username_valid ) {
 				global $blog_id;
 				$current_user = wp_get_current_user();
-				$this->_apply_pseudonym( $current_user->ID, $blog_id, $username, $first_name, $last_name );
-
-				// Get information on the applied pseudonym
-				$blog_url = home_url();
-				$new_info = array(
-					__( 'First Name', 'classblogs' ) => esc_html( $first_name ),
-					__( 'Last Name', 'classblogs' ) => esc_html( $last_name ),
-					__( 'Username', 'classblogs' ) => esc_html( $username ),
-					__( 'Blog URL', 'classblogs' ) => sprintf( '<a href="%1$s">%1$s</a>', esc_url( $blog_url ) ) );
+				$this->_apply_pseudonym( $current_user->ID, $blog_id, $username );
 
 				// Display the updated information to the user
-				$message = array( __( 'You are now blogging using a pseudonym.  Your new user information is as follows.', 'classblogs' ), '<p>' );
-				foreach ( $new_info as $key => $value ) {
+				$blog_url = home_url();
+				$message = array( __( 'You successfully changed your username.  Your new user information is as follows.', 'classblogs' ), '<p>' );
+				$message[] = sprintf( '<strong>%s</strong><br />%s<br /><br />',
+					__( 'Username', 'classblogs'),
+					esc_html( $username ) );
+				if ( ClassBlogs_Utils::is_multisite() ) {
 					$message[] = sprintf( '<strong>%s</strong><br />%s<br /><br />',
-						$key, $value );
+						__( 'Blog URL', 'classblogs'),
+						sprintf( '<a href="%1$s">%1$s</a>', esc_url( $blog_url ) ) );
 				}
 				$message[] = '</p>';
 				ClassBlogs_Admin::show_admin_message( implode( "\n", $message ) );
@@ -123,54 +128,44 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 	<div class="wrap">
 
 		<div id="icon-users" class="icon32"></div>
-		<h2><?php _e( 'Use a Pseudonym', 'classblogs' ); ?></h2>
-
-		<h3><?php _e( 'User Information', 'classblogs' ); ?></h3>
+		<h2><?php _e( 'Change Username', 'classblogs' ); ?></h2>
 
 		<p id="student-pseudonym-instructions">
-			<?php _e( 'You can use the form below to change how you appear on the blog if you do not wish for your real name to be associated with the content that you create.', 'classblogs' ); ?>
+			<?php _e( 'If you have already changed your display name but still wish for there to be no trace of your actual identity on the blog, you can use this page to change the username that you use to log in to the blog.', 'classblogs' ); ?>
 		</p>
 
-		<form method="post" action="">
+		<form method="post" action="" id="cb-username-form">
 
 			<table class="form-table">
 				<tr valign="top">
 					<th scope="row"><?php _e( 'New Username', 'classblogs' ); ?></th>
 					<td>
-						<input type="text" name="new_username" id="new-username" value="<?php echo esc_attr( $username );  ?>" /><br />
+						<input type="text" name="new_username" id="new-username" /><br />
 						<label for="new-username"><?php _e( 'Changing this will change the username that you use to access your blog and the URL at which it can be found.', 'classblogs' ); ?></label>
 						<hr />
 						<label for="new-username">
-							<?php printf( __( 'Your current username is %s', 'classblogs' ), '<strong>' . esc_html( $username ) . '</strong>' ); ?>
+							<?php printf( __( 'Your current username is %s', 'classblogs' ), '<strong>' . esc_html( $current_user->user_login ) . '</strong>' ); ?>
 						</label><br />
-						<label for="new-username">
-							<?php printf( __( 'Your current blog URL is %s', 'classblogs' ), sprintf( '<a href="%1$s">%1$s</a>', esc_url( $blog_url ) ) ); ?>
-						</label><br />
-					</td>
-				</tr>
-				<tr valign="top">
-					<th scope="row"><?php _e( 'New First Name', 'classblogs' ); ?></th>
-					<td>
-						<input type="text" name="new_first_name" id="new-first-name" value="<?php echo esc_attr( $first_name );  ?>"/><br />
-						<label for="new-first-name">
-							<?php printf( __( 'Your current first name is %s', 'classblogs' ), '<strong>' . esc_html( $first_name ) . '</strong>' ); ?>
-						</label>
-					</td>
-				</tr>
-				<tr valign="top">
-					<th scope="row"><?php _e( 'New Last Name', 'classblogs' ); ?></th>
-					<td>
-						<input type="text" name="new_last_name" id="new-last-name" value="<?php echo esc_attr( $last_name );  ?>" /><br />
-						<label for="new-last-name">
-							<?php printf( __( 'Your current last name is %s', 'classblogs' ), '<strong>' . esc_html( $last_name ) . '</strong>' ); ?>
-						</label>
+						<?php if ( ClassBlogs_Utils::is_multisite() ): ?>
+							<label for="new-username">
+								<?php printf( __( 'Your current blog URL is %s', 'classblogs' ), sprintf( '<a href="%1$s">%1$s</a>', esc_url( $blog_url ) ) ); ?>
+							</label>
+						<?php endif; ?>
 					</td>
 				</tr>
 			</table>
 
 			<?php wp_nonce_field( $this->get_uid() ); ?>
-			<p class="submit"><input type="submit" name="Submit" value="<?php _e( 'Use this Pseudonym', 'classblogs' ); ?>" /></p>
+			<p class="submit"><input class="button-primary" type="submit" name="Submit" value="<?php _e( 'Change my Username', 'classblogs' ); ?>" /></p>
+			<strong style="color: #a00">You will only be able to change your username once</strong>
+
 		</form>
+
+		<script type="text/javascript">
+			jQuery("#cb-username-form").submit(function() {
+				return confirm( "<?php _e( 'You can only change your username once.  Are you sure that you wish to change it now?', 'classblogs' ); ?>" );
+			});
+		</script>
 
 	</div>
 <?php
@@ -192,20 +187,13 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 			return false;
 		}
 
-		// If the user name will not change, skip validation
-		if ( $username === wp_get_current_user()->user_login ) {
-			return true;
-		}
-
 		// Check for conflicts with other users and blogs
 		global $wpdb;
-		$valid = validate_username( $username );
-		if ( $valid ) {
+		$valid = validate_username( $username ) && ! username_exists( $username );
+		if ( $valid && ClassBlogs_Utils::is_multisite() ) {
 			$site = get_current_site();
 			$new_path = trailingslashit( $site->path . $username );
-			if ( ClassBlogs_Utils::is_multisite() ) {
-				$valid = ! $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(*) FROM $wpdb->blogs WHERE path=%s", $new_path ) );
-			}
+			$valid = ! $wpdb->get_var( $wpdb->prepare( "SELECT COUNT(*) FROM $wpdb->blogs WHERE path=%s", $new_path ) );
 		}
 		return $valid;
 	}
@@ -223,58 +211,55 @@ class ClassBlogs_Plugins_StudentPseudonym extends ClassBlogs_BasePlugin
 	 * @param  int    $user_id        the student's user ID
 	 * @param  int    $blog_id        the ID of the student's primary blog
 	 * @param  string $new_username   the student's new username
-	 * @param  string $new_first_name the student's new first name
-	 * @param  string $new_last_name  the student's new last name
 	 *
 	 * @access private
 	 * @since 0.1
 	 */
-	private function _apply_pseudonym( $user_id, $blog_id, $new_username, $new_first_name, $new_last_name )
+	private function _apply_pseudonym( $user_id, $blog_id, $new_username )
 	{
 		global $wpdb;
-		ClassBlogs_WordPress::switch_to_blog( $blog_id );
-		$old_url = trailingslashit( home_url() );
 
-		// Update the student's username and their display-name information
-		$full_name = "$new_first_name $new_last_name";
-		update_user_meta( $user_id, 'first_name', $new_first_name );
-		update_user_meta( $user_id, 'last_name', $new_last_name );
-		update_user_meta( $user_id, 'nickname', $new_first_name );
-		wp_update_user( array(
-			'ID' => $user_id,
-			'user_nicename' => str_replace( ' ', '-', $new_username ),
-			'display_name' =>  $full_name ) );
+		// Update the student's username
 		$wpdb->update(
 			$wpdb->users,
 			array( 'user_login' => $new_username ),
 			array( 'ID' => $user_id ),
 			array( '%s' ),
 			array( '%d' ) );
-		update_option( 'blogname', $full_name );
 
-		// Update the student's blog URL
-		$site = get_current_site();
-		$new_path = trailingslashit( $site->path . $new_username );
-		$new_url = 'http://' . $site->domain . $new_path;
-		update_option( 'siteurl', $new_url );
-		update_option( 'home', $new_url );
-		update_blog_details( $blog_id, array( 'path' => $new_path ) );
-		delete_option( 'rewrite_rules' );
+		// Deal with the implications of the updated username in multisite
+		if ( ClassBlogs_Utils::is_multisite() ) {
+			ClassBlogs_WordPress::switch_to_blog( $blog_id );
+			$old_url = trailingslashit( home_url() );
 
-		// Replace any occurrences of the old URL in the blog's posts
-		$referring_posts = $wpdb->get_results( "
-			SELECT ID, post_content FROM $wpdb->posts
-			WHERE post_content LIKE '%%" . like_escape( $old_url ) . "%%' " );
-		foreach ( $referring_posts as $post ) {
-			$wpdb->update(
-				$wpdb->posts,
-				array( 'post_content' => str_replace( $old_url, $new_url, $post->post_content ) ),
-				array( 'ID' => $post->ID ),
-				array( '%s' ),
-				array( '%d' ) );
+			// Update the blog URL to reflect their new username
+			$site = get_current_site();
+			$new_path = trailingslashit( $site->path . $new_username );
+			$new_url = 'http://' . $site->domain . $new_path;
+			update_option( 'siteurl', $new_url );
+			update_option( 'home', $new_url );
+			update_blog_details( $blog_id, array( 'path' => $new_path ) );
+			delete_option( 'rewrite_rules' );
+
+			// Replace any occurrences of the old URL in the blog's posts
+			$referring_posts = $wpdb->get_results( "
+				SELECT ID, post_content FROM $wpdb->posts
+				WHERE post_content LIKE '%%" . like_escape( $old_url ) . "%%' " );
+			foreach ( $referring_posts as $post ) {
+				$wpdb->update(
+					$wpdb->posts,
+					array( 'post_content' => str_replace( $old_url, $new_url, $post->post_content ) ),
+					array( 'ID' => $post->ID ),
+					array( '%s' ),
+					array( '%d' ) );
+			}
+			ClassBlogs_WordPress::restore_current_blog();
 		}
 
-		ClassBlogs_WordPress::restore_current_blog();
+		// Flag that the user has changed their username
+		$changed = $this->get_option( 'changed_users' );
+		$changed[$new_username] = true;
+		$this->update_option( 'changed_users', $changed );
 	}
 }
